Broke out collisions to new files. Added collisionrecord class HeadingIE is now an unsigned short Fixed tracking of tileoffsetXY in seg028_2941_385()
