+++ src/physics/motion_init.cs
@@ -20,7 +20,7 @@ namespace Underworld
             MotionParams.unk_16 = (short)commonObjDat.unk6_5678(itemid);
             MotionParams.scaleresistances_1C = (byte)commonObjDat.scaleresistances(itemid);
             MotionParams.unk_1d = 0;
-            MotionParams.heading_1E = (short)(projectile.heading<<0xD);
+            MotionParams.heading_1E = (ushort)(projectile.heading<<0xD);
             MotionParams.unk_24 = 0;
             MotionParams.radius_22 = (byte)commonObjDat.radius(itemid);
             MotionParams.height_23 = (byte)commonObjDat.height(itemid);
@@ -41,7 +41,7 @@ namespace Underworld
             {
                 MotionParams.x_0 += (short)(projectile.npc_xhome<<3);
                 MotionParams.y_2 += (short)(projectile.npc_yhome<<3);
-                MotionParams.heading_1E = (short)(projectile.ProjectileHeading<<8);
+                MotionParams.heading_1E = (ushort)(projectile.ProjectileHeading<<8);
                 MotionParams.unk_25_tilestate = (sbyte)(1 << projectile.UnkBit_0XA_Bit456);
                 MotionParams.pitch_13 = (sbyte)((projectile.Projectile_Pitch - 16) << 6);
                 MotionParams.unk_10 = (short)(projectile.UnkBit_0X13_Bit7 * -4);
